Implement citation sorting
